% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_files.R
\name{download_data_file}
\alias{download_data_file}
\title{Download data.gov file}
\usage{
download_data_file(url)
}
\arguments{
\item{url}{URL of the file to be downloaded}
}
\description{
Fetch file and save in /data
}
