% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_income.R
\name{tidy_income}
\alias{tidy_income}
\title{Tidy Income XLS}
\usage{
tidy_income(income_sheet)
}
\arguments{
\item{income_sheet}{Total Income sheet from a london borough income data .xls file.}
}
\description{
Turn data.gov london income by borough sheet into a tidy data format
}
